refactor(vs-agent): move demo under examples, seed as a script, filter e2e by filename
