+++ .gitignore
@@ -24,4 +24,3 @@ CLAUDE.md
 
 # Local demo secrets (keep the .example template)
 .env
-.env.demo
