+++ apps/vs-agent/examples/vt-flow-demo/verana-node-init.sh
