+++ apps/vs-agent/vitest.config.ts
@@ -1,6 +1,6 @@
 import { existsSync } from 'fs'
 /* eslint-disable import/no-unresolved */
-import { defineConfig, mergeConfig } from 'vitest/config'
+import { configDefaults, defineConfig, mergeConfig } from 'vitest/config'
 
 import rootConfig from '../../vitest.config'
 
@@ -11,6 +11,7 @@ export default mergeConfig(
     test: {
       setupFiles: existsSync(setup) ? [setup] : [],
       include: ['tests/**/*.test.ts'],
+      exclude: [...configDefaults.exclude, '**/*.e2e.test.ts'],
       reporters: ['verbose'],
       outputFile: './test-results.json',
       // askar-nodejs native bindings race when multiple test files
